+++ src/Tests/TEST_Workbench/Workbench.cs
@@ -7,7 +7,7 @@ using System.Threading.Tasks;
 using Machina;
 using Machina.Types.Geometry;
 using Machina.Descriptors.Models;
-
+using Machina.Solvers.Errors;
 
 namespace TEST_Workbench
 {
@@ -22,63 +22,23 @@ namespace TEST_Workbench
             Machina.Logger.WriteLine += Console.WriteLine;
 
             //List<double> target = new List<double> { 0, 0, 0, 0, 90, 0 };
-            List<double> target = new List<double> { 72.474, 67.689, -126.868, 0, -30.821, 287.526 };
+            List<double> target = new List<double> { 259, 67.689, -126.868, 0, -30.821, 287.526 };
+
+            List<SolverError> errors;
 
             robotModel = RobotModel.CreateABBIRB140();
-            var frames = robotModel.ForwardKinematics(target, Units.Degrees);
+            var frames = robotModel.ForwardKinematics(target, Units.Degrees, out errors);
 
-            //var it = 0;
-            //foreach (var m in frames)
-            //{
-            //    Console.WriteLine(it);
-            //    Console.WriteLine(m);
-            //    it++;
-            //}
+            Console.WriteLine(frames.Last());
 
-            Console.WriteLine(Plane.CreateFromMatrix(frames[frames.Count - 1]));
-
-            bot = Robot.Create("FKTest", "ABB");
-            bot.ConnectionManager(ConnectionType.Machina);
-            bot.ControlMode(ControlType.Online);
-            bot.Connect();
-
-            bot.SolutionFKReceived += Arm_SolutionFKReceived;
-
-            bot.Message("FK TEST STARTING");
-
-            for (int i = 0; i < 25; i++)
+            foreach (var err in errors)
             {
-                Axes a = Axes.RandomFromDoubles(-400, 400);
-                string msg = "20 " + a.ToWhitespacedValues();
-                bot.CustomCode(msg);
+                Console.WriteLine(err);
             }
-
-            Console.WriteLine("Press any key to DISCONNECT...");
-            Console.ReadKey();
-
-            bot.Disconnect();
 
             Console.WriteLine("Press any key to EXIT...");
             Console.ReadKey();
         }
-
-        private static void Arm_SolutionFKReceived(object sender, Machina.EventArgs.SolutionFKReceivedArgs args)
-        {
-            var frames = robotModel.ForwardKinematics(args.Axes.ToList(), Units.Degrees);
-            var tcp = frames.Last();
-            Console.WriteLine($"Robot: {args.TCP.ToArrayString(6)}\n   FK: {tcp.ToArrayString(6)}");
-            Console.WriteLine($"SIMILAR: {args.TCP.IsSimilarTo(tcp, 0.001)}");
-
-        }
-
-        static void Scale(List<double> rots, int factor)
-        {
-            for (int i = 0; i < rots.Count; i++)
-            {
-                rots[i] *= factor;
-            }
-        }
-
 
 
     }